Changed dtype of mention_idxs to long

+++ blink/biencoder/data_process_mult.py
@@ -193,7 +193,7 @@ def process_mention_data(
     n_labels = torch.tensor(
         select_field(processed_samples, "n_labels"), dtype=torch.int,
     )
-    mention_idx = torch.arange(len(n_labels), dtype=torch.int)
+    mention_idx = torch.arange(len(n_labels), dtype=torch.long)
 
     tensor_data = TensorDataset(context_vecs, label_idxs, n_labels, mention_idx)
